+++ src/fedefs/hdiv_bdm1.jl
@@ -201,7 +201,7 @@ function get_coefficients(::Type{ON_CELLS}, FE::FESpace{Tv, Ti, <:HDIVBDM1, APT}
     xCellFaceSigns::Union{VariableTargetAdjacency{Int32}, Array{Int32, 2}} = xgrid[CellFaceSigns]
     nfaces::Int = num_faces(EG)
     dim::Int = dim_element(EG)
-    return function closure(coefficients::Array{<:Real, 2}, cell::Int)
+    return function closure(coefficients::Array{<:Real, 2}, cell)
         fill!(coefficients, 1.0)
         # multiplication with normal vector signs (only RT0)
         for j in 1:nfaces, k in 1:dim
@@ -216,7 +216,7 @@ function get_coefficients(::Type{ON_CELLS}, FE::FESpace{Tv, Ti, <:HDIVBDM1, APT}
     xCellFaceSigns::Union{VariableTargetAdjacency{Int32}, Array{Int32, 2}} = xgrid[CellFaceSigns]
     nfaces::Int = num_faces(EG)
     dim::Int = dim_element(EG)
-    return function closure(coefficients::Array{<:Real, 2}, cell::Int)
+    return function closure(coefficients::Array{<:Real, 2}, cell)
         fill!(coefficients, 1.0)
         for j in 1:nfaces, k in 1:dim
             coefficients[k, 3 * j - 2] = xCellFaceSigns[j, cell] # RT0
